Add utilities for common VAS interactions
* Added a `Client.py` module that acts as a central hub for all
communication with VAS.  This type ensures that proper HTTP headers are
set on requests as well as ensuring the that resulting HTTP status codes
indicate successful results.  In the case that the status code indicates
an unsuccessful result this type raises an error indicating what has
gone wrong with the request.  In the case that a request is task-based,
the call blocks until the task has completed.

* Added a `LinkUtils.py` module that parses links from the standard
payload link structure.  Understands how to navigate both top-level
links and the `self` links that are nested within the payloads of
collections.