Normalize property names with Ruby API
In a recent commit, a number of property names were changed to include their _group_ or _node_ -ness.  This caused the API to diverge from the Ruby API and the examples to break as noted in Issue #1.

Most of the changed property names have been returned to their original names and now match the Ruby API.

This commit fixes Issue #1.