Apply permissions after agent completely unpacked
The vFabric Administration Agent file permissions are very locked down. Because of this, applying permissions incrementally as each directory and file are unpacked does not work.  Because the directories do not have write permissions, files cannot be unpacked into directories.

To remedy this in `AgentImage.extract_to()`, all directories and files are unpacked with umask permissions first and then the zip file permissions are applied to each path in a second pass.  The permissions are retrieved and cached during the unpack phase so only a single read pass of the zip file is required.