Make implementation match Ruby implementation
In the interests of maintainability, the implementation of the Python language binding has been updated so that it matches the implementation of the Ruby language binding to the maximum extent possible.  This means that the implementations have the same module and class structure, and the arguments to each function are now in the same order.  In addition, the documentation has been updated to match that of the Ruby language binding.