Add `__str__` implementations
In the interests of making the API more usable in both a debugger and a console, implementations of `__str__` have been added to all non-collection types.  These implementations print the name of the class and the values of the important properties for each type.  _Note_ that not all property values are included in the output as that would make many of them nearly unreadable.